+++ tests/phpunit/testcases/Session/SessionInitTest.php
@@ -47,11 +47,13 @@ class SessionInitTest extends \WP_UnitTestCase
         $directory = self::$files_directory["session"]["files"];
         $ignoreFiles = ['.gitignore', '.', '..'];
 
-        $files = scandir($directory);
+        if(file_exists($directory)){
+            $files = scandir($directory);
 
-        foreach ($files as $file) {
-            if (!in_array($file, $ignoreFiles)) {
-                unlink($directory . '/' . $file);
+            foreach ($files as $file) {
+                if (!in_array($file, $ignoreFiles)) {
+                    unlink($directory . '/' . $file);
+                }
             }
         }
     }
